requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
babel==2.16.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg==3.2.3
psycopg-binary==3.2.3
py-moneyed==3.0
pydotplus==2.0.2
pygraphviz==1.13
PyJWT==2.9.0
pyparsing==3.1.4
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
redis==5.1.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
